хз шото там исправлял

+++ lib/data_base.rb
@@ -1,6 +1,5 @@
 module DataBase
   class Base
-    prepend BaseInterface
 
     attr_reader :client
 
@@ -8,16 +7,24 @@ module DataBase
       @client = client
     end
 
-    def set_data(key, data)
-      client.set(key.to_s, data.to_json)
+    def set_data(table, key, data)
+      client.hset(table, key, data.to_json)
     end
 
-    def get_data(key)
-      new_data(client.get(key.to_s)).to_h
+    def get_data(table, key)
+      new_data(client.hget(table, key)).to_h
     end
 
-    def clear_data(key)
-      client.set(key.to_s, "{}")
+    def get_hash_all_data(table)
+      client.hgetall(table)
+    end
+
+    def clear_data(table, key)
+      client.del(table, key.to_s)
+    end
+
+    def clear_table(table)
+      client.del(table)
     end
 
     def new_data(data)
@@ -25,17 +32,16 @@ module DataBase
     end
 
     def delete_all
-      client.del(*client.keys)
+      client.del(*client.keys) unless client.keys.empty?
     end
   end
 
   class Data
-    prepend DataInterface
 
     attr_reader :value
 
     def initialize(data)
-      @value = (data&.empty? or data.nil?) ? '{}' : data
+      @value = (data&.empty? or data.nil?) ? "{}" : data
     end
 
     def to_h
